# server/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Backend",
        "version": "0.1.0"
    },
    "paths": {
...
        "/evaluate/{prompt_id}": {
            "post": {
                "tags": [
                    "evaluate"
                ],
                "summary": "Evaluate",
                "operationId": "evaluate",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "prompt_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Prompt Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EvaluateSummary"
                                }
                            }
                        }
                    },
                    "409": {
                        "detail": "document not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Conflict"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "db"
                ],
                "summary": "Post User",
                "operationId": "post_user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRootInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "detail": "document already exists"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{username}": {
            "get": {
                "tags": [
                    "db"
                ],
                "summary": "Get User",
                "operationId": "get_user",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Email of the user to retrieve",
                            "title": "Username"
                        },
                        "description": "Email of the user to retrieve"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "detail": "document not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Assertion": {
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/BaseAssertionTypes"
                    },
                    "value": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Value"
                    },
                    "threshold": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Threshold"
                    },
                    "weight": {
                        "anyOf": [
                            {
                                "type": "number",
                                "minimum": 1.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Weight"
                    },
                    "metric": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Metric"
                    }
                },
                "type": "object",
                "required": [
                    "type"
                ],
                "title": "Assertion"
            },
            "BaseAssertionTypes": {
                "type": "string",
                "enum": [
                    "contains",
                    "contains-all",
                    "contains-any",
                    "contains-json",
                    "contains-sql",
                    "contains-xml",
                    "cost",
                    "equals",
                    "icontains",
                    "icontains-all",
                    "icontains-any",
                    "is-json",
                    "is-sql",
                    "is-xml",
                    "javascript",
                    "latency",
                    "levenshtein",
                    "perplexity-score",
                    "perplexity",
                    "python",
                    "regex",
                    "rouge-n",
                    "starts-with"
                ],
                "title": "BaseAssertionTypes"
            },
...
            "EvaluateResult": {
                "properties": {
                    "response": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ProviderResponse"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "vars": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "type": "string"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Vars"
                    },
                    "score": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Score"
                    },
                    "latencyMs": {
                        "type": "integer",
                        "title": "Latencyms"
                    },
                    "gradingResult": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/GradingResult"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "namedScores": {
                        "additionalProperties": {
                            "type": "number"
                        },
                        "type": "object",
                        "title": "Namedscores"
                    },
                    "cost": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cost"
                    },
                    "metadata": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Metadata"
                    }
                },
                "type": "object",
                "required": [
                    "score",
                    "latencyMs",
                    "namedScores"
                ],
                "title": "EvaluateResult"
            },
            "EvaluateStats": {
                "properties": {
                    "successes": {
                        "type": "integer",
                        "title": "Successes"
                    },
                    "failures": {
                        "type": "integer",
                        "title": "Failures"
                    },
                    "tokenUsage": {
                        "$ref": "#/components/schemas/TokenUsage"
                    }
                },
                "type": "object",
                "required": [
                    "successes",
                    "failures",
                    "tokenUsage"
                ],
                "title": "EvaluateStats"
            },
            "EvaluateSummary": {
                "properties": {
                    "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Timestamp"
                    },
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/EvaluateResult"
                        },
                        "type": "array",
                        "title": "Results"
                    },
                    "stats": {
                        "$ref": "#/components/schemas/EvaluateStats"
                    }
                },
                "type": "object",
                "required": [
                    "timestamp",
                    "results",
                    "stats"
                ],
                "title": "EvaluateSummary"
            },
            "GradingResult": {
                "properties": {
                    "pass": {
                        "type": "boolean",
                        "title": "Pass"
                    },
                    "score": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Score"
                    },
                    "componentResults": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/GradingResult"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Componentresults"
                    },
                    "assertion": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Assertion"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "comment": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Comment"
                    }
                },
                "type": "object",
                "required": [
                    "pass",
                    "score"
                ],
                "title": "GradingResult"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "HttpExceptionModel": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "status": {
                        "type": "integer",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "status"
                ],
                "title": "HttpExceptionModel"
            },
...
            "ProviderResponse": {
                "properties": {
                    "cached": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cached"
                    },
                    "cost": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cost"
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error"
                    },
                    "logProbs": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "number"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Logprobs"
                    },
                    "metadata": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Metadata"
                    },
                    "output": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Output"
                    },
                    "tokenUsage": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/TokenUsage"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "title": "ProviderResponse"
            },
...
            "SuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "Success"
                    }
                },
                "type": "object",
                "title": "SuccessResponse"
            },
...
            "TokenUsage": {
                "properties": {
                    "total": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Total",
                        "default": 0
                    },
                    "prompt": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Prompt",
                        "default": 0
                    },
                    "completion": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Completion",
                        "default": 0
                    },
                    "cached": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cached",
                        "default": 0
                    }
                },
                "type": "object",
                "title": "TokenUsage"
            },
            "User": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "_id": {
                        "type": "string",
                        "title": " Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "email"
                ],
                "title": "User"
            },
            "UserRootInput": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "email"
                ],
                "title": "UserRootInput"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/auth/login"
                    }
                }
            }
        }
    }
}